requirements.txt
slack_sdk
slack_bolt
python-dotenv
boto3